Move auth_provider logic into own file
